chore: tree-graph to extend pkg
